day123/cloudwatch-collector/requirements.txt
# AWS SDK
boto3==1.34.110
botocore==1.34.110

# Web Framework
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6

# Async Support
aiohttp==3.9.5
asyncio==3.4.3

# Data Processing
python-json-logger==2.0.7
pydantic==2.7.1

# Metrics and Monitoring
prometheus-client==0.20.0

# Storage and Caching
redis==5.0.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
moto==5.0.6

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
colorama==0.4.6
requests==2.31.0

# Development
black==24.4.2
pylint==3.1.0
